cleanup(runtime): remove legacy duckdb database module
