--- packages/runtime/src/_legacy/database/src/index.ts
@@ -1,6 +1,0 @@
-/**
- * Entry point for @obora/database.
- * Public API only.
- */
-
-export * from "./public-api.js";
